# docs/swagger.yaml
definitions:
...
  models.ResActivityList:
    properties:
      activityId:
        type: integer
      airdropEndTime:
        type: integer
      airdropNumber:
        type: integer
      airdropStartTime:
        type: integer
      chainId:
        type: integer
      claimedPercentage:
        type: number
      creator:
        type: string
      daoAddress:
        type: string
      endTime:
        type: integer
      price:
        type: string
      publishTime:
        type: integer
      stakingAmount:
        type: string
      startTime:
        type: integer
      title:
        type: string
      tokenAddress:
        type: string
      tokenChainId:
        type: integer
      types:
        type: string
    type: object
  models.ResActivityPage:
    properties:
      list:
        items:
          $ref: '#/definitions/models.ResActivityList'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /stpdao/v2/activity/list:
    get:
      description: query activity list
      parameters:
      - description: chainId
        in: query
        name: chainId
        type: integer
      - description: Dao address
        in: query
        name: daoAddress
        type: string
      - description: status:Soon→Open→Ended→Airdrop→Closed
        in: query
        name: status
        type: string
      - description: types:Airdrop,PublicSale
        in: query
        name: types
        type: string
      - description: offset,page
        in: query
        name: offset
        required: true
        type: integer
      - description: count,page
        in: query
        name: count
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResActivityPage'
      summary: query activity list
      tags:
      - activity
...
swagger: "2.0"